backend/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# PDF Generation
weasyprint==63.1
jinja2==3.1.4

# Geospatial (flight path maps + area calc)
shapely==2.0.6
pyproj==3.7.0
staticmap==0.5.7
folium==0.18.0

# HTTP Client (for OpenDroneLog + Ollama)
httpx==0.28.1

# Email
aiosmtplib==3.0.2

# Utilities
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
Pillow==11.0.0
